GraphsAsScenarios component
