+++ src/components/GraphsAsScenarios/GraphsAsScenarios.module.css
@@ -1,0 +1,7 @@
+ul.GraphsAsScenarios {
+    /**/
+    outline: 1px dotted rgb(255, 38, 38);
+    /**/
+
+
+}
